Abstract Library Service in an interface and implements it in two services
